requirements.txt
Flask==3.1.0
adafruit-circuitpython-ds3231==2.4.23
adafruit-circuitpython-fram==1.3.24
ntplib==0.4.0
RPi.GPIO==0.7.1
Nuitka==2.6